+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Minor optimizations to the Archiver.
 - `cat` command now accepts a prefix ID for blobs and trees.
 - Added `completion` command to generate autocompletion scripts.
+- Restore symlink metadata on Unix.
 
 # v0.1.0-beta.2a
 ### Changes